requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
pymupdf==1.23.8
unstructured==0.11.8
openai>=1.6.0
httpx>=0.25.0
faiss-cpu==1.7.4
numpy==1.24.3
python-dotenv==1.0.0
streamlit==1.28.1
requests==2.31.0
sentence-transformers==2.2.2
scikit-learn==1.3.2 